# Reuse allocated buffer space in merge iterator. (#9538)